+++ db/migrations/0039_narrow_klaw.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "plataforma"."planes" ADD COLUMN "meses_bonificados" integer DEFAULT 0 NOT NULL;